PhaseFromInterferograms

PhaseFromInterferograms.FineTilts

PhaseFromInterferograms.PSIAlg

PhaseFromInterferograms.RoughTilts

PhaseFromInterferograms.TiltExtractionAlg

PhaseFromInterferograms.Windowing.GaussianWindow

PhaseFromInterferograms.zoomFFT2D.FFT2Zoom

PhaseFromInterferograms.zoomFFT2D.FFT2Zoom

PhaseFromInterferograms.FindHarmonics.eraseZerothOrder!

PhaseFromInterferograms.FindHarmonics.findfineharmonic

PhaseFromInterferograms.FindHarmonics.findroughharmonic

PhaseFromInterferograms.fourier_tilt

PhaseFromInterferograms.get_LS_phase_from_n_psi

PhaseFromInterferograms.get_diff_phase_from_n_psi

PhaseFromInterferograms.get_phase_from_n_psi

PhaseFromInterferograms.get_tilt

PhaseFromInterferograms.getfinetilt

PhaseFromInterferograms.gettilts

PhaseFromInterferograms.zoomFFT2D.expsquared

PhaseFromInterferograms.zoomFFT2D.findfirstharmonic2

PhaseFromInterferograms.zoomFFT2D.removeDC

PhaseFromInterferograms.FineTilts — Type

 FineTilts <: TiltExtractionAlg

Method of extracting tilts from an array of the interferogram differences based on locating maxima in the spectra with subpixel accuracy using zoomFFT (aka CZT or Bluestein algorithm).

Parameters:

erasesize = 2 – half-diameter of the central lobe to be excluded: block with size 2 erasesize + 1 centered around the origin

cropsize = 2 – half diameter of the side lobe used tolocate the maximum: block with size 2 cropsize + 1 centerd around the maximum of the spectrum.

zoomlevels = nothing – array of zoomlevels used for iterational subpixel approximation. E.g. zoomlevels = [1,4] will first find the side lobe maximum in the original Fourier spectrum , and then in the spectrum sampled with 4 times higher rate. If set to nothing uses automatic sequence of the zoom levels.

normals = [nothing] – array of the normals defining the half plane where the maximum is located.

PhaseFromInterferograms.PSIAlg — Type

PSIAlg <: AbstractAlg

An abstract algorithm for Phase-shifting Interferometry (PSI).

Concrete instances are callable or can be used in get_phase_from_n_psi functions:

`(alg::PSIAlg)(measurements, deltas)`

extracts the phase ϕ from several measurements Iₖ = b + a cos(ϕ + δₖ).

deltas is an array of δₖ and it should be the same length as the number of measurements or one less. In this case, δ₁ is asuumed to be zero.

See also get_phase_from_n_psi](@ref).

PhaseFromInterferograms.RoughTilts — Type

 RoughTilts <: TiltExtractionAlg

Simple method of extracting approximate tilts from an array of the interferogram differences based on the Fourier filtering.

Parameters:

erasesize = 2 – half-diameter of the central lobe to be excluded: block with size 2 erasesize + 1 centered around the origin

selectsize = 2 – half diameter of the side lobe used to reconstruct the tilt: block with size 2 selectsize + 1 centerd around the maximum of the spectrum.

PhaseFromInterferograms.TiltExtractionAlg — Type

TiltExtractionAlg <: AbstractAlg

Abstract algorithm for retrieval tilts from interferogram differences.

Concrete instances of the type are callable:

`(alg::TiltExtractionAlg)(idiffs)` -> tilts

or can be used in gettilts function.

See also gettilts.

PhaseFromInterferograms.fourier_tilt — Function

fourier_tilt(τ, σ, arrsize, fftshift = false)

Construct an array of values of linear function t(x) = τ⋅x +σ compatible with Fourier transform coordinates, that is T(ξ)=F(exp(i t(x))) has maximum at ξ=τ/(2π). Coordinates: ξ is defined by fftfreq, x defined with the origin at arrsize÷2+1 for fftshift = true and with the origin at at the first element of the array if fftshift = false.

PhaseFromInterferograms.get_LS_phase_from_n_psi — Method

get_LS_ phase_from_n_psi(images, deltas)

Given n interferometric images with known phase shifts deltas between them, reconstruct the phase using the least-squares diversity-based method.

PhaseFromInterferograms.get_diff_phase_from_n_psi — Method

get_diff_phase_from_n_psi(images, deltas)

Given n+1 interferometric images with n known phase shifts deltas between them, reconstruct the phase using the phase-difference-based approach.

PhaseFromInterferograms.get_phase_from_n_psi — Method

get_phase_from_n_psi(images, deltas, alg::PSIAlg; kwargs...)

Retreive phase from several interferograms obtained with phase-shifting interferometry.

PhaseFromInterferograms.get_tilt — Function

get_tilt(idiff, erasesize=3, selectsize=3, debug=false)

Get the approximate tilt from the interferogram difference through the inverse Fourier of the part of the Fourier spectrum of idiff loacated around the position of the maximal side lobe.

Parameters:

erasesize – minimal distance from the sidelobe to the DC component

selectsize – number of the neigbour Fourier components to include in the tilt reconstruction.

PhaseFromInterferograms.getfinetilt — Method

getfinetilt(idiff; n=[1, 0], zoomlevels=nothing, visualdebug=false, erasesize=2, cropsize=2) -> tilt, τ, σ

TBW

PhaseFromInterferograms.gettilts — Method

gettilts(idiffs, alg::TiltExtractionAlg)

Extract tilts from the interferogram differences idiffs using alg method. Return tilts and additional information depending on the method used.

PhaseFromInterferograms.Windowing.GaussianWindow — Type

GaussianWindow(w) is an object corresponding to a centerd gaussian window  `w` pixels wide.

`w` can be a single number or a tuple, with specified width for each dimension.

PhaseFromInterferograms.FindHarmonics.eraseZerothOrder! — Method

eraseZerothOrder(spectrum, method = "block"; erasesize = 3)

Erase the zeroth order in the spectrum. Currently impemented methods:

"block" – set to zero a cube of width 2*erasesize +1 around the DC frequency.

PhaseFromInterferograms.FindHarmonics.findfineharmonic — Method

findfineharmonic(array, h, params)

Document this function

PhaseFromInterferograms.FindHarmonics.findroughharmonic — Method

findroughharmonic(array, params)

Document this function

PhaseFromInterferograms.zoomFFT2D.FFT2Zoom — Method

FFT2Zoom(M,N,R,S,αx, αy)

Structure containing necessary arrays for calculation of zoomed 2D FFT

PhaseFromInterferograms.zoomFFT2D.FFT2Zoom — Method

(f::FFT2Zoom)(x)

Calculate zoomed FFT of x

PhaseFromInterferograms.zoomFFT2D.expsquared — Method

expsquared(x, coef)

Calculate exp(coef * x^2)

PhaseFromInterferograms.zoomFFT2D.findfirstharmonic2 — Method

findfirstharmonic2(a, zoomlevels = nothing) ->(freqs, amp), amp_hist, freqs_hist

Find first harmonic in a 2D array a with subpixel precision, using sequential FFT2Zoom with zoomlevels

PhaseFromInterferograms.zoomFFT2D.removeDC — Function

removeDC(a, size = 2)

Remove the zeroth lobe of width 2*size+1 from the signal